+++ src/note.rs
@@ -37,14 +37,24 @@ impl RandomSeed {
         }
     }
 
-    /// from bytes
+//<<<<<<< HEAD
+//    /// from bytes
+//=======
+    /// Reads a note's random seed from bytes, given the note's nullifier.
+    ///
+    /// Returns `None` if the nullifier is not for the same note as the seed.
+//>>>>>>> d05b6cee9df7c4019509e2f54899b5979fb641b5
     pub fn from_bytes(rseed: [u8; 32], rho: &Nullifier) -> CtOption<Self> {
         let rseed = RandomSeed(rseed);
         let esk = rseed.esk_inner(rho);
         CtOption::new(rseed, esk.is_some())
     }
 
-    /// as bytes
+//<<<<<<< HEAD
+//    /// as bytes
+//=======
+    /// Returns the byte array corresponding to this seed.
+//>>>>>>> d05b6cee9df7c4019509e2f54899b5979fb641b5
     pub fn as_bytes(&self) -> &[u8; 32] {
         &self.0
     }
@@ -124,7 +134,25 @@ impl PartialEq for Note {
 impl Eq for Note {}
 
 impl Note {
-    pub(crate) fn from_parts(
+//<<<<<<< HEAD
+//    pub(crate) fn from_parts(
+//=======
+    /// Creates a `Note` from its component parts.
+    ///
+    /// Returns `None` if a valid [`NoteCommitment`] cannot be derived from the note.
+    ///
+    /// # Caveats
+    ///
+    /// This low-level constructor enforces that the provided arguments produce an
+    /// internally valid `Note`. However, it allows notes to be constructed in a way that
+    /// violates required security checks for note decryption, as specified in
+    /// [Section 4.19] of the Zcash Protocol Specification. Users of this constructor
+    /// should only call it with note components that have been fully validated by
+    /// decrypting a received note according to [Section 4.19].
+    ///
+    /// [Section 4.19]: https://zips.z.cash/protocol/protocol.pdf#saplingandorchardinband
+    pub fn from_parts(
+//>>>>>>> d05b6cee9df7c4019509e2f54899b5979fb641b5
         header: u64,
         recipient: Address,
         d1: NoteValue,
@@ -134,8 +162,13 @@ impl Note {
         rho: Nullifier,
         rseed: RandomSeed,
         memo: [u8; 512]
-    ) -> Self {
-        Note {
+//<<<<<<< HEAD
+//    ) -> Self {
+//        Note {
+//=======
+    ) -> CtOption<Self> {
+        let note = Note {
+//>>>>>>> d05b6cee9df7c4019509e2f54899b5979fb641b5
             header,
             recipient,
             d1,
@@ -145,7 +178,12 @@ impl Note {
             rho,
             rseed,
             memo
-        }
+//<<<<<<< HEAD
+//        }
+//=======
+        };
+        CtOption::new(note, note.commitment_inner().is_some())
+//>>>>>>> d05b6cee9df7c4019509e2f54899b5979fb641b5
     }
 
     /// Generates a new note.
@@ -165,6 +203,7 @@ impl Note {
         memo: [u8; 512]
     ) -> Self {
         loop {
+/*<<<<<<< HEAD
             let note = Note {
                 header,
                 recipient,
@@ -178,6 +217,11 @@ impl Note {
             };
             if note.commitment_inner().is_some().into() {
                 break note;
+=======*/
+            let note = Note::from_parts(header, recipient, d1, d2, sc, nft, rho, RandomSeed::random(&mut rng, &rho), memo);
+            if note.is_some().into() {
+                break note.unwrap();
+//>>>>>>> d05b6cee9df7c4019509e2f54899b5979fb641b5
             }
         }
     }